SimpleDeviceConfig: Add initial implementation
This is a simple app that provisions the DeviceConfig settings namespace
on boot with values defined in a resource overlay. It is meant to be
delegated as the device configurator in the framework config
(config_deviceConfiguratorPackageName) to make sure it receives the
correct permissions.

This can be useful for e.g. configuring proprietary Google services
without delegating GMS as the device configurator to prevent it from
overriding custom default settings, or minimizing changes to core
framework code by making changes in DeviceConfig instead as intended by
Google.

Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>